+++ implementation/BusinessLogic/src/main/java/businesslogic/FlightManagerImpl.java
@@ -6,6 +6,7 @@ import businessentitiesapi.Flight;
 import businessentitiesapi.FlightManager;
 import persistence.FlightStorageService;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalDateTime;
@@ -36,7 +37,7 @@ public class FlightManagerImpl implements FlightManager {
             LocalDateTime depTime,
             LocalDateTime arrTime,
             String airplane,
-            int basePrice
+            BigDecimal basePrice
     ){
         return new Flight(
                 flightID,
